refactor: 회원가입 디렉토리 이동 #8
